Added ApplicationListener and CommandLineRunner

+++ src/main/java/com/stackroute/service/MuzixServiceImpl.java
@@ -3,6 +3,7 @@ package com.stackroute.service;
 
         import com.stackroute.domain.Muzix;
         import com.stackroute.exceptions.MuzixAlreadyExistsException;
+        import com.stackroute.exceptions.MuzixTrackNotFoundException;
         import com.stackroute.repository.MuzixRepository;
         import org.springframework.beans.factory.annotation.Autowired;
         import org.springframework.stereotype.Service;
@@ -14,6 +15,7 @@ public class MuzixServiceImpl implements MuzixService
 {
 
     MuzixRepository muzixRepository;
+    int a=9;
 
     //constructor
 
@@ -51,15 +53,23 @@ public class MuzixServiceImpl implements MuzixService
 
     @Override
     public Muzix updateMuzix(Muzix muzix) {
+
         return muzixRepository.save(muzix);
     }
 
     // Deleting the track by ID
     @Override
-    public Muzix removeMuzix(int trackId) {
-        Muzix deleteMuzix = muzixRepository.getOne(trackId);
-        muzixRepository.delete(deleteMuzix);
-        return deleteMuzix;
-    }
+    public Muzix removeMuzix(int trackId) throws MuzixTrackNotFoundException{
+        if (!muzixRepository.existsById(trackId)) {
+            throw new MuzixTrackNotFoundException("Sorry!The Track is Empty");
+        }
+            Muzix deleteMuzix = muzixRepository.getOne(trackId);
+            muzixRepository.delete(deleteMuzix);
 
+            return deleteMuzix;
+        }
+        @Override
+    public List<Muzix> trackByName(String trackName) {
+        return muzixRepository.findBytrackName(trackName);
+    }
 }
